backend/Auth/requirements.txt
flask
flask-cors
pymongo
python-dotenv
werkzeug
dnspython

